Added in More Functionality
Can now make a move, check for moves, empty log file, show a move
which includes writing to the correct file.

Adding in proper support for setupBoard() so that it will get in
the enemy's move, parse it, make sure it's a valid move, and then
update the board with that move.